# add index to node-authorizer for high cardinality vertices