@liggitt 2018-05-02
@mikedanese comments addressed

@wojtek-t - would be interested in seeing a profile captured from a run on this PR to compare to the one mentioned in https://github.com/kubernetes/kubernetes/issues/62808#issuecomment-382651517. is there a PR-equivalent of that job we can run?