@liggitt 2018-05-02
/hold
I wanted to add some more unit tests to verify index invalidation on update