@liggitt 2018-05-02
/hold cancel

added an index invalidation test