@wojtek-t 2018-05-03
> @wojtek-t - would be interested in seeing a profile captured from a run on this PR to compare to the one mentioned in #62808 (comment). is there a PR-equivalent of that job we can run?

Given it's already merged, we should have some results from upcoming postsubmits. I will let you know (probably on Monday when back from kubecon).