@liggitt 2018-05-06
Comparison of run 148 and 149

Before:
```
      "data": {
        "Perc50": 0.11,
        "Perc90": 0.163,
        "Perc99": 1.428
      },
      "unit": "ms",
      "labels": {
        "Count": "10415010",
        "Resource": "secrets",
        "Scope": "namespace",
        "Subresource": "",
        "Verb": "GET"
      }
```

After, cut 99th percentile in half:
```
      data": {
        "Perc50": 0.101,
        "Perc90": 0.153,
        "Perc99": 0.708
      },
      "unit": "ms",
      "labels": {
        "Count": "10314765",
        "Resource": "secrets",
        "Scope": "namespace",
        "Subresource": "",
        "Verb": "GET"
      }
``` 


Sample CPU profiles before:
* https://storage.googleapis.com/kubernetes-jenkins/logs/ci-kubernetes-e2e-gce-scale-performance/148/artifacts/profiles/ApiserverCPUProfile_2018-05-02T09:12:19Z.pdf
* https://storage.googleapis.com/kubernetes-jenkins/logs/ci-kubernetes-e2e-gce-scale-performance/148/artifacts/profiles/ApiserverCPUProfile_2018-05-02T10:07:50Z.pdf

Sample CPU profiles after:
* https://storage.googleapis.com/kubernetes-jenkins/logs/ci-kubernetes-e2e-gce-scale-performance/149/artifacts/profiles/ApiserverCPUProfile_2018-05-04T09:15:14Z.pdf
* https://storage.googleapis.com/kubernetes-jenkins/logs/ci-kubernetes-e2e-gce-scale-performance/149/artifacts/profiles/ApiserverCPUProfile_2018-05-04T10:10:45Z.pdf


Node authorization dropped from 4-12% before to not being present in the profile after. The index-building code path did not appear in the CPU profile either

Will keep an eye on memory use trends over the next few runs to make sure the index memory use impact wasn’t significant